+++ lib/fundamentals/networking/main.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_training/fundamentals/networking/ui/posts_screen.dart';
+
+void main() {
+  runApp(
+    MaterialApp(
+      home: SafeArea(
+        child: Container(
+          color: Colors.white,
+          child: Material(child: PostsScreen()),
+        ),
+      ),
+    ),
+  );
+}